Criterion A Documentation - Planning
The first criterion required by an IB IA is the basic planning of the project. You can find more information about it here: http://ib.compscihub.net/wp-content/uploads/2015/04/IBCompSciGuide.pdf (p82), but the essence of the mark scheme is that the project should appropriately address the issues raised by the client with good reason to be developed.