Completed the deletion and addition!
Completed the commandLine project. Some of the code, is quite dodgy, BUT, it works well. Currently, only working with the two people that are on the API, but when they get added to the API everything will work automatically! To my (and Gautam's) employers, this code is a stepping stone on our paths to fluency in Java, not the endpoint. I hope you hire us regardless. :)